Group by search

+++ back-end/src/main/java/com/eldes/devsuperior/dsvendas/dtos/SaleDto.java
@@ -11,7 +11,7 @@ public class SaleDto implements Serializable {
 	private Long id;
 	private Integer visited;
 	private Integer deals;
-	private Integer amount;
+	private Double amount;
 	private LocalDate date;
 
 	private SellerDto seller;
@@ -20,7 +20,7 @@ public class SaleDto implements Serializable {
 	public SaleDto() {}
 
 
-	public SaleDto(Long id, Integer visited, Integer deals, Integer amount, LocalDate date, SellerDto seller) {
+	public SaleDto(Long id, Integer visited, Integer deals, Double amount, LocalDate date, SellerDto seller) {
 		this.id = id;
 		this.visited = visited;
 		this.deals = deals;
@@ -63,11 +63,11 @@ public class SaleDto implements Serializable {
 		this.deals = deals;
 	}
 
-	public Integer getAmount() {
+	public Double getAmount() {
 		return this.amount;
 	}
 
-	public void setAmount(Integer amount) {
+	public void setAmount(Double amount) {
 		this.amount = amount;
 	}
